Update: check ghế trước khi lưu, xóa ghế, tài khoản dùng ajax,

+++ TicketLand_project/Areas/Admin/Controllers/manage_membersController.cs
@@ -130,8 +130,6 @@ namespace TicketLand_project.Areas.Admin.Controllers
         }
 
         // POST: Admin/manage_members/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "member_id,member_name,username,password,gender,date_of_birth,email,city,phone,role_id,member_point")] member member, HttpPostedFileBase imageFile)
@@ -156,20 +154,7 @@ namespace TicketLand_project.Areas.Admin.Controllers
             return View(member);
         }
 
-        // GET: Admin/manage_members/Delete/5
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            member member = db.members.Find(id);
-            if (member == null)
-            {
-                return HttpNotFound();
-            }
-            return View(member);
-        }
+ 
 
         // POST: Admin/manage_members/Delete/5
         [HttpPost, ActionName("Delete")]
@@ -178,6 +163,10 @@ namespace TicketLand_project.Areas.Admin.Controllers
         {
 
             member member = db.members.Find(id);
+            if (member == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             db.members.Remove(member);
             db.SaveChanges();
             return RedirectToAction("Index");
